refactor: entity 파일 infrastructure로 이동
