+++ mbti_service/src/provider/mysql.provider.ts
@@ -3,6 +3,8 @@ import { SnakeNamingStrategy } from 'typeorm-naming-strategies';
 import { addTransactionalDataSource } from 'typeorm-transactional';
 import { ConfigType } from '@nestjs/config';
 import databaseConfig from '../config/database.config';
+import { Test } from '../test/infrastructure/entity/test.entity';
+import { Question } from '../test/infrastructure/entity/question.entity';
 
 //todo : develop 과 main 에 대한 환경변수 별도로 설정
 export const mysqlProvider = {
@@ -17,7 +19,7 @@ export const mysqlProvider = {
       username: config.database.user,
       password: config.database.password,
       database: config.database.db,
-      entities: [__dirname + '/test/domain/*{.ts,.js}'],
+      entities: [Test, Question],
       namingStrategy: new SnakeNamingStrategy(),
       logging: true,
       synchronize: config.database.synchronize,
