+++ mbti_service/src/test/infrastructure/entity/question.entity.ts
